requirement.txt
django 2.2
djangocrispyforms
